implement comment function

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -21,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -40,7 +40,7 @@ public class PublishController {
     public String doPublish(@RequestParam(name="title", required = false) String title,
                             @RequestParam(name="description", required = false) String description,
                             @RequestParam(name="tag", required = false) String tag,
-                            @RequestParam(name="id", required = false) Integer id,
+                            @RequestParam(name="id", required = false) Long id,
                             HttpServletRequest request,
                             Model model) {
         model.addAttribute("title", title);
